Adding required step
