+++ dartpad_codelabs/src/null_safety_workshop/step_05/solution.dart
@@ -1,12 +1,24 @@
+void printThreeValues({
+  required int first,
+  required int second,
+  required int third,
+}) {
+  print('First is $first, second is $second, '
+  	  'and third is $third.');
+}
+
+void printFourValues({
+  int? first,
+  required int second,
+  int? third,
+  required int fourth
+}) {
+  print('First is $first, second is $second, '
+  	  'third is $third, and fourth is $fourth.');
+}
+
 void main() {
-  String text;
-
-  if (DateTime.now().hour < 12) {
-    text = "It's morning! Let's make aloo paratha!";
-  } else {
-    text = "It's afternoon! Let's make biryani!";
-  }
-
-  print(text);
-  print(text.length);
+  printThreeValues(first: 1, second: 2, third: 3);
+  printFourValues(first: 1, second: 2, third: 3,
+  	  fourth: 4);
 }
