+++ dartpad_codelabs/src/null_safety_workshop/step_06/instructions.md
@@ -1,7 +1,13 @@
-# Type promotion via null checks
-Dart can also promote nullable variables to non-nullable ones when there's a
-null check in the same scope.
+# Type promotion
+With sound null safety, Dart’s flow analysis has been extended to take
+nullability into account. Nullable variables that can’t possibly contain null
+values are treated like non-nullable variables. This behavior is called type
+promotion.
+
+Dart’s type system can track where variables are assigned and read, and can
+verify that non-nullable fields are given values before any code tries to read
+from them. This process is called definite assignment.
 
 ## Exercise
-In the code below, add an `if` statement to the beginning of `getLength` that
-returns zero if `str` is null:
+Try uncommenting the if-else statement in the code below, and watch the analyzer
+errors disappear:
