+++ dartpad_codelabs/src/null_safety_workshop/step_06/snippet.dart
@@ -1,9 +1,12 @@
-int getLength(String? str) {
-  // Add null check here
-
-  return str.length;
-}
-
 void main() {
-  print(getLength('This is a string!'));
+  String text;
+
+  // if (DateTime.now().hour < 12) {
+  //  text = "It's morning! Let's make aloo paratha!";
+  // } else {
+  //  text = "It's afternoon! Let's make biryani!";
+  // }
+
+  print(text);
+  print(text.length);
 }
