+++ dartpad_codelabs/src/null_safety_workshop/step_07/instructions.md
@@ -1,7 +1,7 @@
-# Promotion via exceptions
-Promotion works with exceptions as well as return statements. If you check a
-nullable variable for null and throw an exception, Dart will consider any future
-reads from that variable in the same scope to be null-safe.
+# Type promotion via null checks
+Dart can also promote nullable variables to non-nullable ones when there's a
+null check in the same scope.
 
 ## Exercise
-Try a null check that throws an Exception instead of returning zero.
+In the code below, add an `if` statement to the beginning of `getLength` that
+returns zero if `str` is null:
