+++ dartpad_codelabs/src/null_safety_workshop/step_08/instructions.md
@@ -1,13 +1,7 @@
-# Object properties aren't promotable
-A common "Huh?" moment when first learning to use Dart's null safety arrives the
-first time you do a null check on an object field. This is due to the fact that
-in Dart, object properties have getters and setters (either real or implied),
-and so something that looks like a typical value property can return different
-things when read twice in the same code block.
+# Promotion via exceptions
+Promotion works with exceptions as well as return statements. If you check a
+nullable variable for null and throw an exception, Dart will consider any future
+reads from that variable in the same scope to be null-safe.
 
-Fortunately, there's a simple trick to handle this: read the variable once and
-cache it in a local variable@
-
-## Exercise: 
-Try adding a local variable to store the value returned by `provider.value` in
-the example, then check and print that variable instead.
+## Exercise
+Try a null check that throws an Exception instead of returning zero.
