+++ dartpad_codelabs/src/null_safety_workshop/step_09/solution.dart
@@ -1,14 +1,6 @@
 import 'dart:math';
 
-class StringProvider {
-  String? value = 'A String!';
-}
-
-class RandomStringProvider extends StringProvider {
-  @override
-  void set value(String? v) {}
-
-  @override
+class RandomStringProvider {
   String? get value =>
     Random().nextBool() ? 'A String!' : null;
 }
@@ -16,7 +8,7 @@ class RandomStringProvider extends StringProvider {
 void printString(String str) => print(str);
 
 void main() {
-  StringProvider provider = RandomStringProvider();
+  final provider = RandomStringProvider();
 
   final str = provider.value;
