+++ dartpad_codelabs/src/null_safety_workshop/step_10/snippet.dart
@@ -1,13 +1,26 @@
-class Meal {
-  String description;
+import 'dart:math';
 
-  void setDescription(String str) {
-    description = str;
-  }
+class StringProvider {
+  String? value = 'A String!';
 }
 
+class RandomStringProvider extends StringProvider {
+  @override
+  void set value(String? v) {}
+
+  @override
+  String? get value =>
+    Random().nextBool() ? 'A String!' : null;
+}
+
+void printString(String str) => print(str);
+
 void main() {
-  final myMeal = Meal();
-  myMeal.setDescription('Feijoada!');
-  print(myMeal.description);
+  StringProvider provider = RandomStringProvider();
+
+  if (provider.value == null) {
+    print('The value is null.');
+  }
+
+  printString(provider.value);
 }
